Better first paint + identicon for users.
When I added fontawesome I used the svg library which is javascript. This then rendered an SVG after the DOM loaded. Because fontawesome had to wait, the icons would appear later. Also because the icons were SVG, they were not affected by the font-size of the element. Long story short, I switched back to css font-awesome so the icons are cached and loaded very fast.